pyproject.toml
[project]
name = "meli-challenge"
version = "0.1.0"
description = "MLOps Challenge - Housing Price Prediction"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "ipykernel>=7.1.0",
    "joblib>=1.3.0",
    "matplotlib>=3.10.8",
    "mlflow>=2.17.0",
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "prometheus-client>=0.21.0",
    "pydantic-settings>=2.0.0",
    "scikit-learn>=1.8.0",
    "uvicorn>=0.34.0",
    "xgboost>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
]

[project.scripts]
train = "train:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
]
